==> Vector-DB/06.Multimodal_rec/04. Image_input_module.ipynb <==
from PIL import Image
from transformers import YolosFeatureExtractor, YolosForObjectDetection
import matplotlib.pyplot as plt
import numpy as np
from torchvision.transforms import ToTensor

import json
import pandas as pd
import warnings
import openai
import os
from dotenv import load_dotenv
load_dotenv()

# Ignore warnings
warnings.filterwarnings('ignore')
# pandas dataframe display
pd.set_option('display.max_columns', None)

# initialize openai
openai.api_key = os.environ["OPENAI_API_KEY"]

### 1. yolo를 활용한 object detection

attributes = pd.read_csv("../data/imaterialist-fashion-2020-fgvc7/attribute_specific.csv")

from yolo_utils import fix_channels, visualize_predictions, rescale_bboxes, plot_results, box_cxcywh_to_xyxy

MODEL_NAME = "valentinafeve/yolos-fashionpedia"

feature_extractor = YolosFeatureExtractor.from_pretrained('hustvl/yolos-small')
model = YolosForObjectDetection.from_pretrained(MODEL_NAME)

# 미리 선정된 prediction labels
cats = ['shirt, blouse', 'top, t-shirt, sweatshirt', 'sweater', 'cardigan', 'jacket', 'vest', 'pants', 'shorts', 'skirt', 'coat', 'dress', 'jumpsuit', 'cape', 'glasses', 'hat', 'headband, head covering, hair accessory', 'tie', 'glove', 'watch', 'belt', 'leg warmer', 'tights, stockings', 'sock', 'shoe', 'bag, wallet', 'scarf', 'umbrella', 'hood', 'collar', 'lapel', 'epaulette', 'sleeve', 'pocket', 'neckline', 'buckle', 'zipper', 'applique', 'bead', 'bow', 'flower', 'fringe', 'ribbon', 'rivet', 'ruffle', 'sequin', 'tassel']

IMAGE_PATH = '../data/imaterialist-fashion-2020-fgvc7/test_images/test_image5.jpg'

image = Image.open(open(IMAGE_PATH, "rb"))
image = fix_channels(ToTensor()(image))
# image

inputs = feature_extractor(images=image, return_tensors="pt")
outputs = model(**inputs)

# visualize_predictions(image, outputs, threshold=0.) # threshold가 높을수록 안전

probas = outputs.logits.softmax(-1)[0, :, :-1]
len(probas)

probas = outputs.logits.softmax(-1)[0, :, :-1]
keep = probas.max(-1).values > 0.5

prob = probas[keep]

len(prob)

indices = [np.argmax(idx.detach().numpy()) for idx in prob]

# indices

detected_cats = [cats[idx] for idx in indices]

detected_cats

len(outputs.pred_boxes[0])

boxes = outputs.pred_boxes[0, keep].cpu()

# boxes

bboxes_scaled = rescale_bboxes(boxes, image.size).tolist()

# bboxes_scaled

COLORS = [[0.000, 0.447, 0.741], [0.850, 0.325, 0.098], [0.929, 0.694, 0.125],
          [0.494, 0.184, 0.556], [0.466, 0.674, 0.188], [0.301, 0.745, 0.933]]

def plot_results_2(pil_img, labels, boxes):
    plt.figure(figsize=(16,10))
    plt.imshow(pil_img)
    ax = plt.gca()
    colors = COLORS * 100
    for label, (xmin, ymin, xmax, ymax), c in zip(labels, boxes, colors):
        ax.add_patch(plt.Rectangle((xmin, ymin), xmax - xmin, ymax - ymin,
                                   fill=False, color=c, linewidth=3))
        
        ax.text(xmin, ymin, label, fontsize=10,
                bbox=dict(facecolor=c, alpha=0.8))
    plt.axis('off')
    plt.show()

# plot_results_2(image, detected_cats, bboxes_scaled)

### filter bounding boxes (필요한 카테고리만 선별)

new_df = pd.read_csv("../data/imaterialist-fashion-2020-fgvc7/clothes_final2.csv")
new_df.name.unique()

category_of_interest = new_df.name.unique().tolist()

keep_indices = list()
keep_bboxes = list()

for idx, box in zip(detected_cats, bboxes_scaled):
    if idx in category_of_interest:
        keep_indices.append(idx)
        keep_bboxes.append(box)

keep_indices, keep_bboxes

### 2. bbox 통합

def iou(boxA, boxB):
    # Calculate the intersection rectangle
    xA = max(boxA[0], boxB[0])
    yA = max(boxA[1], boxB[1])
    xB = min(boxA[2], boxB[2])
    yB = min(boxA[3], boxB[3])

    # Compute the area of intersection
    interArea = max(0, xB - xA) * max(0, yB - yA)

    # Compute the area of both the prediction and ground-truth rectangles
    boxAArea = (boxA[2] - boxA[0]) * (boxA[3] - boxA[1])
    boxBArea = (boxB[2] - boxB[0]) * (boxB[3] - boxB[1])

    # Compute the intersection over union by taking the intersection
    # area and dividing it by the sum of prediction + ground-truth
    # areas - the intersection area
    iou = interArea / float(boxAArea + boxBArea - interArea)

    return iou

def merge_boxes(boxes, labels):
    merged_boxes = []
    merged_labels = []
    used = set()

    for i in range(len(boxes)):
        if i in used:
            continue
        current_box = boxes[i]
        for j in range(i + 1, len(boxes)):
            if j in used or labels[i] != labels[j]:
                continue
            if iou(current_box, boxes[j]) > 0.5:  # Assuming a positive IoU indicates overlap
                # For xyxy format, we merge by finding the min and max coordinates
                current_box = [
                    min(current_box[0], boxes[j][0]), 
                    min(current_box[1], boxes[j][1]), 
                    max(current_box[2], boxes[j][2]), 
                    max(current_box[3], boxes[j][3])
                ]
                used.add(j)
        merged_boxes.append(current_box)
        merged_labels.append(labels[i])
        used.add(i)

    return np.array(merged_boxes), merged_labels


# plot_results_2(image, keep_indices, keep_bboxes)

merged_bbox, merged_labels = merge_boxes(keep_bboxes, keep_indices)

merged_labels, merged_bbox

# plot_results_2(image, merged_labels, merged_bbox.tolist())

### 3. cropping

from image_utils import crop_bbox

from PIL import Image, ImageFilter

def resize_img(image, category):
    standard_size = {"lowerbody":[420, 540],
        "upperbody":[500, 700],
        "wholebody":[480, 880],
        "legs and feet":[100, 150],
        "head":[150, 100],
        "others":[200, 350],
        "waist":[200, 100],
        "arms and hands":[75, 75],
        "neck":[120, 200]}
    
    w, h = image.size
    img_size = w*h

    new_width, new_height = standard_size[category]
    new_size = new_width * new_height

    if img_size >= new_size:
        # For downsizing
        downsized_image = image.resize((new_width, new_height), Image.Resampling.LANCZOS)
        return downsized_image
    else:
        # For upsizing
        upsized_image = image.resize((new_width, new_height), Image.Resampling.LANCZOS)
        upsized_image = upsized_image.filter(ImageFilter.UnsharpMask(radius=2, percent=150, threshold=3))
        return upsized_image

categories = new_df[['supercategory', 'name']].drop_duplicates()

# 여기서 dictionary로 변환을 하면서 각 카테고리를 키 값으로 설정해준다
# 따라서 '신발'이 두 개 detect 되더라도 하나만 선택
cropped_images = dict()

for label, box in zip(merged_labels, merged_bbox):
    cropped = resize_img(crop_bbox(image, box), categories.loc[categories['name']==label, 'supercategory'].values[0])
    cropped_images[label] = cropped

cropped_images

### 4. 서치 가능한 이미지 여부 판단

from pinecone import Pinecone

pc_api_key = os.environ["PINECONE_API_KEY"]
pc = Pinecone(api_key=pc_api_key)
# index 개수 확인
# index_list = pc.list_indexes().indexes

# index description
index = pc.Index("fastcampus")
index.describe_index_stats()

# CLIP
from image_utils import fetch_clip, extract_img_features, draw_images

model, processor, tokenizer = fetch_clip(model_name="patrickjohncyh/fashion-clip")

cropped_images

results = dict()

for label, image in cropped_images.items():
    img_emb = extract_img_features(image, processor, model).tolist()

    result = index.query(
        vector=img_emb[0],
        top_k=5,
        filter={"category": {"$eq": label}},
        include_metadata=True
    )

    paths = [i['metadata']['img_path'] for i in result.matches]

    results[label] = paths


# for k, paths in results.items():
#     print(k)
#     draw_images([Image.open('../data/'+i) for i in paths])

def clothes_detector(image, feature_extractor, model, thresh=0.5):
    # all categories
    # Yolo-V8의 Categories
    cats = ['shirt, blouse', 'top, t-shirt, sweatshirt', 'sweater', 'cardigan', 'jacket', 'vest', 'pants', 'shorts', 
            'skirt', 'coat', 'dress', 'jumpsuit', 'cape', 'glasses', 'hat', 'headband, head covering, hair accessory', 
            'tie', 'glove', 'watch', 'belt', 'leg warmer', 'tights, stockings', 'sock', 'shoe', 'bag, wallet', 'scarf', 
            'umbrella', 'hood', 'collar', 'lapel', 'epaulette', 'sleeve', 'pocket', 'neckline', 'buckle', 'zipper', 
            'applique', 'bead', 'bow', 'flower', 'fringe', 'ribbon', 'rivet', 'ruffle', 'sequin', 'tassel']
    # category we are interested in
    category_of_interest = ['pants', 'shirt, blouse', 'jacket', 'top, t-shirt, sweatshirt', 'dress', 'shoe', 'glasses', 
                        'skirt', 'bag, wallet', 'belt', 'headband, head covering, hair accessory', 'sock', 'hat', 
                        'watch', 'glove', 'tights, stockings', 'sweater', 'tie', 'shorts', 'scarf', 'coat', 'vest', 
                        'umbrella', 'cardigan', 'cape', 'jumpsuit', 'leg warmer']
    # yolo detection
    inputs = feature_extractor(images=image, return_tensors="pt")
    outputs = model(**inputs)
    
    # extract detected labels and boundingboxes
    probas = outputs.logits.softmax(-1)[0, :, :-1]
    keep = probas.max(-1).values > thresh

    prob = probas[keep]

    indices = [np.argmax(idx.detach().numpy()) for idx in prob]
    detected_cats = [cats[idx] for idx in indices]
    boxes = outputs.pred_boxes[0, keep].cpu()

    bboxes_scaled = rescale_bboxes(boxes, image.size).tolist()
    
    # keep boxes that we are interested in
    keep_indices = list()
    keep_bboxes = list()

    for idx, box in zip(detected_cats, bboxes_scaled):
        if idx in category_of_interest:
            keep_indices.append(idx)
            keep_bboxes.append(box)
    # overlapping한 구간이 있는 bbox들을 통합
    merged_bbox, merged_labels = merge_boxes(keep_bboxes, keep_indices)

    # cropping
    categories = pd.read_csv("../data/imaterialist-fashion-2020-fgvc7/categories.csv")
    cropped_images = dict()

    for label, box in zip(merged_labels, merged_bbox):
        cropped = resize_img(crop_bbox(image, box), categories.loc[categories['name']==label, 'supercategory'].values[0])
        cropped_images[label] = cropped

    return cropped_images

def image_search(index, cropped_images, model, processor, top_k=10):
    results = dict()

    for label, image in cropped_images.items():
        img_emb = extract_img_features(image, processor, model).tolist()

        result = index.query(
            vector=img_emb[0],
            top_k=top_k,
            filter={"category": {"$eq": label}},
            include_metadata=True
        )

        results[label] = result
    return results

### 5. 발견한 각 카테고리 별 서치 결과 제공

MODEL_NAME = "valentinafeve/yolos-fashionpedia"

feature_extractor = YolosFeatureExtractor.from_pretrained('hustvl/yolos-small')
model = YolosForObjectDetection.from_pretrained(MODEL_NAME)

clip_model, clip_processor, clip_tokenizer = fetch_clip(model_name="patrickjohncyh/fashion-clip")

- 패션과 전혀 상관없는 강아지 사진을 사용하여 테스트
- 결과가 빈 dict로 나오는 것이 정상

image = Image.open("../data/imaterialist-fashion-2020-fgvc7/test_images/test.jpg")
image = fix_channels(ToTensor()(image))
# image

cropped_items = clothes_detector(image, feature_extractor, model, thresh=0.5)

cropped_items

search_result = image_search(index, cropped_items, clip_model, clip_processor)

# 이미지들의 path 가져오기
paths = dict()
for k,v in search_result.items():
    paths[k] = ['../data/'+i['metadata']['img_path'] for i in v['matches']]

# 이미지들 show
for k,v in paths.items():
    print(k)
    draw_images([Image.open(i) for i in v])
