==> Vector-DB/06.Multimodal_rec/05. Hybrid_search_module.ipynb <==
import json
import pandas as pd
from tqdm.notebook import tqdm
import base64
import requests
from PIL import Image
import openai
import os
import numpy as np
from dotenv import load_dotenv

load_dotenv()
pd.set_option('display.max_columns', None)

### 1. GPT4v를 이용하여 우리들이 갖고 있는 attribute에 맞게 특징 추출 -> 텍스트로 변환
#### Enrich image with descriptions

# initialize openai
openai.api_key = os.environ["OPENAI_API_KEY"]

```python
"'main_category', 'silhouette', 'silhouette_fit', 'waistline',
       'length', 'collar_type', 'neckline_type', 'sleeve_type',
       'pocket_type', 'opening_type', 'non-textile material type',
       'leather', 'textile finishing, manufacturing techniques',
       'textile pattern', 'animal', 'other'"
```

from search_utils import clothes_detector
from transformers import YolosFeatureExtractor, YolosForObjectDetection

MODEL_NAME = "valentinafeve/yolos-fashionpedia"

feature_extractor = YolosFeatureExtractor.from_pretrained('hustvl/yolos-small')
model = YolosForObjectDetection.from_pretrained(MODEL_NAME)

image = Image.open("../data/imaterialist-fashion-2020-fgvc7/test_images/test_image5.jpg").convert("RGB")

cropped_images = clothes_detector(image, feature_extractor, model)

cropped_images

cropped_images['shoe']

def encode_image(image_path):
  with open(image_path, "rb") as image_file:
    return base64.b64encode(image_file.read()).decode('utf-8')

import io
import base64

# GPT를 이용해서 이미지에 대한 description 생성

def describe_clothes(image, label, openai_key):
  buffer = io.BytesIO()
  # Save the image to the buffer in JPEG format
  image.save(buffer, format="JPEG")
  buffer.seek(0)
  image_data = buffer.read()

  base64_image = base64.b64encode(image_data).decode('utf-8')
  image_desc_prompt = """Focus on {} inside the image.
        Identify the attributes of the item.
        The attributes you should answer are : 
        - clothes_type
        - color
        - silhouette
        - silhouette_fit
        - waisteline
        - sleeve_type
        - collar_type
        - length
        - gender
        - patterns
        - textile_pattern
        
        Ignore the attributes you cannot answer.
        Keep the answer simple and clear, having max three words per attribute.
  """.format(label)

  headers = {
    "Content-Type": "application/json",
    "Authorization": f"Bearer {openai_key}"
  }

  payload = {
    "model": "gpt-4-vision-preview",
    "messages": [
      {
        "role": "user",
        "content": [
          {
            "type": "text",
            "text": image_desc_prompt
          },
          {
            "type": "image_url",
            "image_url": {
              "url": f"data:image/jpeg;base64,{base64_image}"
            }
          }
        ]
      }
    ],
    "max_tokens": 300
  }

  response = requests.post("https://api.openai.com/v1/chat/completions", headers=headers, json=payload)

  return response.json()['choices'][0]['message']['content']

cropped_images.items()

descriptions = dict()

for i, img in cropped_images.items():
    print(i)
    desc = describe_clothes(img, i, openai.api_key)
    descriptions[i] = desc

descriptions

### 2. Text search : top-100 (broad search)

from search_utils import fashion_query_transformer, text_search
from image_utils import draw_images

text_query = fashion_query_transformer(str(descriptions))
text_query

from splade.splade.models.transformer_rep import Splade
from transformers import AutoTokenizer

splade_model_id = 'naver/splade-cocondenser-ensembledistil'

splade_model = Splade(splade_model_id, agg='max')
splade_model.to('cpu')
splade_model.eval()

splade_tokenizer = AutoTokenizer.from_pretrained(splade_model_id)

from pinecone import Pinecone

pc_api_key = os.environ["PINECONE_API_KEY"]
pc = Pinecone(api_key=pc_api_key)
# index 개수 확인
# index_list = pc.list_indexes().indexes

# index description
index = pc.Index("fastcampus")
# index.describe_index_stats()

# CLIP
from image_utils import fetch_clip, extract_img_features, draw_images

model, processor, tokenizer = fetch_clip(model_name="patrickjohncyh/fashion-clip")

results = text_search(index, text_query, model, tokenizer, splade_model, splade_tokenizer, top_k=100, hybrid=False)
results.keys()

len(results['bag, wallet'].matches)

paths = dict()
for k,v in results.items():
    paths[k] = ['../data/'+i['metadata']['img_path'] for i in v['matches']]

# 이미지들 show
for k,v in paths.items():
    print(k)
    draw_images([Image.open(i) for i in v[:10]]) # 10개씩만 display

### 3. Image search를 활용한 2차 서치
#### 매칭된 이미지 내에서 한 번 더 서치를 진행 (image search)

# PC index의 제한으로 인해 유사 데이터를 local DB와 같이 사용
local_db = pd.read_csv("../data/imaterialist-fashion-2020-fgvc7/local_db.csv")
local_db['values'] = local_db['values'].apply(json.loads)
local_db.shape

local_db.head(3)

각 카테고리별로 서치된 결과 값을 저장

results.keys()

ids = list()

for category, value in results.items():
    id = [i['id'] for i in value['matches']]
    ids.append({category:id})

ids[0].keys(), list(ids[0].values())[0][:3], "..."

ids[1]['shoe']

# 이미지 embedding
# cropped_images도 필요

final_results = list()

for search_result in ids:
    category = list(search_result.keys())[0]
    search_ids = list(search_result.values())[0]
    # 관련 
    filtered_local_db = local_db.loc[local_db['vdb_id'].isin(search_ids)]
    
    img_emb = extract_img_features(cropped_images[category], processor, model)

    # def search_local_db()
    def calculate_dot_products(embedding, df, column_name):
        dot_products = df[column_name].apply(lambda x: np.dot(embedding, x))
        return dot_products

    # Calculate dot products
    dot_products = calculate_dot_products(img_emb.cpu().numpy()[0], filtered_local_db, 'values')

    # Find the indices of the top 5 most similar embeddings
    top_indices = dot_products.nlargest(10).index

    # Retrieve the top 5 most similar embeddings
    top_similar_ids = filtered_local_db.loc[top_indices, 'vdb_id'].tolist()
    
    final_results.append({category:top_similar_ids})

cropped_images

filtered_local_db

final_results

for search_result in final_results:
    category = list(search_result.keys())[0]
    paths = list(search_result.values())[0]

    full_paths = [os.path.join("../data/imaterialist-fashion-2020-fgvc7", "cropped_images", i+".jpg") for i in paths]
    print(category)
    draw_images([Image.open(i) for i in full_paths])

## Function으로 변환

from search_utils import get_single_text_embedding

def calculate_dot_products(embedding, df, column_name):
    dot_products = df[column_name].apply(lambda x: np.dot(embedding, x))
    return dot_products


def get_top_indices(db, input_data, category, clip_processor, clip_model, clip_tokenizer, top_k, type='image'):
    if type=='image':
        # input_data should be a single cropped image
        emb = extract_img_features(input_data, clip_processor, clip_model)
        # Calculate dot products
        dot_products = calculate_dot_products(emb.cpu().numpy()[0], db, 'values')
    elif type=='text':
        # input_data should be a single string of text
        emb = get_single_text_embedding(input_data, clip_model, clip_tokenizer)
        # Calculate dot products
        dot_products = calculate_dot_products(np.array(emb)[0], db, 'values')

    # Find the indices of the top 5 most similar embeddings
    top_indices = dot_products.nlargest(top_k).index

    # Retrieve the top 5 most similar embeddings
    top_similar_ids = db.loc[top_indices, 'vdb_id'].tolist()

    return {category:top_similar_ids}


def additional_search(local_db, cropped_images, search_results, clip_processor, clip_model, clip_tokenizer, top_k=10):
    
    ids = list()
    for category, value in search_results.items():
        id = [i['id'] for i in value['matches']]
        ids.extend(id)

    final_results = list()

    # 전체 아이템 중, 1차 retrieve 된 것들만 가져옴
    db = local_db.loc[local_db['vdb_id'].isin(ids)]

    for label, v in search_results.items(): # 텍스트로부터
        tmp = db.loc[db['name']==label]

        # 텍스트에도 있고, 이미지에도 레이블이 있는 경우
        if label in cropped_images.keys():
            r = get_top_indices(tmp, cropped_images[label], label, clip_processor, clip_model, clip_tokenizer, top_k, type='image')
            final_results.append(r)
        # 텍스트에만 있는 경우, 그냥 top_k를 가져온다
        else:
            final_results.append({ label : [i['id'] for i in v['matches']][:top_k]} )

    refined_result = dict()

    for search_result in final_results:
        category = list(search_result.keys())[0]
        paths = list(search_result.values())[0]

        full_paths = [os.path.join("../data/imaterialist-fashion-2020-fgvc7", "cropped_images", i+".jpg") for i in paths]
        refined_result[category] = full_paths

        
    return refined_result

## Test

MODEL_NAME = "valentinafeve/yolos-fashionpedia"

feature_extractor = YolosFeatureExtractor.from_pretrained('hustvl/yolos-small')
model = YolosForObjectDetection.from_pretrained(MODEL_NAME)

clip_model, clip_processor, clip_tokenizer = fetch_clip(model_name="patrickjohncyh/fashion-clip")

from search_utils import clothes_detector

image = Image.open("../data/imaterialist-fashion-2020-fgvc7/test_images/test_image4.jpg")
# image = fix_channels(ToTensor()(image))
# image

cropped_images = clothes_detector(image, feature_extractor, model, 0.5)

cropped_images

descriptions = dict()

for i, img in cropped_images.items():
    print(i)
    desc = describe_clothes(img, i, openai.api_key)
    descriptions[i] = desc

for i, v in descriptions.items():
    print(i)
    print(v)
    print()

text_query = fashion_query_transformer(str(descriptions))
text_query

results = text_search(index, text_query, clip_model, clip_tokenizer, splade_model, splade_tokenizer, top_k=100, hybrid=False)
results.keys()

paths = dict()
for k,v in results.items():
    paths[k] = ['../data/'+i['metadata']['img_path'] for i in v['matches']]

# 이미지들 show
for k,v in paths.items():
    print(k)
    draw_images([Image.open(i) for i in v[:10]]) # 10개씩만 display

final_results = additional_search(local_db, cropped_images, results, clip_processor, clip_model, clip_tokenizer)

for k,v in final_results.items():
    print(k)
    draw_images([Image.open(i) for i in v])
