==> Vector-DB/06.Multimodal_rec/08. Fashion_recommender.ipynb <==
import json
import pandas as pd
from tqdm.notebook import tqdm
from PIL import Image
from openai import OpenAI
import openai
import os
from pprint import pprint
from dotenv import load_dotenv
load_dotenv()
pd.set_option('display.max_columns', None)

# initialize openai
openai.api_key = os.environ["OPENAI_API_KEY"]

import base64
import requests

def encode_image(image_path):
  with open(image_path, "rb") as image_file:
    return base64.b64encode(image_file.read()).decode('utf-8')

- 구체적일 수록 좋음
- 하지 말아야 할 것 보다는 해야 할 것 위주로 설명

text_desc = "convert the coat into a gradient colored green coat."

# GPT를 이용해 이미지의 description 생성
# description 최대한 상세하고 구체적으로 작성 필요
# GPT에게 해야할 것 위주로 prompt 작성

image_desc_prompt = """
Analyze the user provided input to create a 'modified description' of the image that reflects the 'user input'.

Incorporating user input, the fashion description should only change slightly from the original image description.
The overall color or the details should be the same unless the user input specifically mention modifications.

The item type should only change if the user input specifically mention modifications.
As an example, if the bomber jacket is in the image, the output should also be a bomber jacket unless the user specifies it to change.

Remember, the total length of your response, including all characters and spaces, must stay within the 500-letter constraint. 
Aim for clarity and brevity in your answer.

#user input : {}
""".format(text_desc)

# Path to your image
image_path = "../data/imaterialist-fashion-2020-fgvc7/test_images/test_image4.jpg"

# Getting the base64 string
base64_image = encode_image(image_path)

headers = {
  "Content-Type": "application/json",
  "Authorization": f"Bearer {openai.api_key}"
}

payload = {
  "model": "gpt-4-vision-preview",
  "messages": [
    {
      "role": "user",
      "content": [
        {
          "type": "text",
          "text": image_desc_prompt
        },
        {
          "type": "image_url",
          "image_url": {
            "url": f"data:image/jpeg;base64,{base64_image}"
          }
        }
      ]
    }
  ],
  "max_tokens": 800
}

response = requests.post("https://api.openai.com/v1/chat/completions", headers=headers, json=payload)


img_desc = response.json()['choices'][0]['message']['content']
pprint(img_desc)

dall-e input

text_prompt = """Create a visual representation of the following fashion description. 
Focus on capturing the essence of the outfit in a realistic manner without overcomplication.
The background should be subtle and not detract from the outfit itself, 
Choose a minimalistic background that complements the style of the attire:
The image should be hyper-realistic

Fashion Description:
{}
""".format(img_desc)

pprint(text_prompt)

client = OpenAI()

response = client.images.generate(
  model="dall-e-3",
  prompt=text_prompt,
  size="1024x1024",
  quality="standard",
  n=1,
  style='vivid'
)

image_url = response.data[0].url
# url is live for 60 seconds after generation
print(image_url)

## Function

def create_modifications(text_desc, image_path):
    # GPT를 이용해서 이미지를 바탕으로 modified description을 생성

    image_desc_prompt = """
    Analyze the user provided input and ensure the description accurately reflects the 'user preference'.

    Incorporating user input, the modified fashion description emphasizes their unique color palette, 
    showcases an updated fashion style with innovative textiles, introduces intricate patterns for visual interest, 
    and highlights a distinctive shape that redefines their overall silhouette.
    It should consider both the fashion in the image, and the 'user input'.

    Remember, the total length of your response, including all characters and spaces, must stay within the 500-letter constraint. 
    Aim for clarity and brevity in your answer.

    #user input : {}
    """.format(text_desc)

    # # Path to your image
    # image_path = "test_images/test_image5.jpg"

    # Getting the base64 string
    base64_image = encode_image(image_path)

    headers = {
    "Content-Type": "application/json",
    "Authorization": f"Bearer {openai.api_key}"
    }

    payload = {
    "model": "gpt-4-vision-preview",
    "messages": [
        {
        "role": "user",
        "content": [
            {
            "type": "text",
            "text": image_desc_prompt
            },
            {
            "type": "image_url",
            "image_url": {
                "url": f"data:image/jpeg;base64,{base64_image}"
            }
            }
        ]
        }
    ],
    "max_tokens": 800
    }

    response = requests.post("https://api.openai.com/v1/chat/completions", headers=headers, json=payload)

    # Dall-E intpu
    
    img_desc = response.json()['choices'][0]['message']['content']
    text_desc = "more formal, suitable for a wedding event. a green coat"

    text_prompt = """Create a visual representation of the following fashion description. 
    Focus on capturing the essence of the outfit in a realistic manner without overcomplication.
    The background should be subtle and not detract from the outfit itself, 
    Choose a minimalistic background that complements the style of the attire:

    Fashion Description:
    {}
    """.format(img_desc)
    

    client = OpenAI()

    response = client.images.generate(
    model="dall-e-3",
    prompt=text_prompt,
    size="1024x1024",
    quality="standard",
    n=1,
    style='vivid'
    )

    image_url = response.data[0].url
    # url is live for 60 seconds after generation
    return image_url
