==> Vector-DB/06.Multimodal_rec/09. User_interaction.ipynb <==
import openai
import os
from PIL import Image
import pandas as pd
import json
from torchvision.transforms import ToTensor
from dotenv import load_dotenv
load_dotenv()

## 0. 필요한 모델 및 function 불러오기

pineconeDB에 연결

from pinecone import Pinecone

# initialize openai
openai.api_key = os.environ["OPENAI_API_KEY"]

pc_api_key = os.environ["PINECONE_API_KEY"]
pc = Pinecone(api_key=pc_api_key)

index = pc.Index("fastcampus")
index.describe_index_stats()

이미지 & 텍스트 dense vector 생성을 위한 CLIP

from image_utils import fetch_clip, draw_images, extract_img_features

clip_model, clip_processor, clip_tokenizer = fetch_clip(model_name="patrickjohncyh/fashion-clip")

text sparse vector 생성을 위한 SPLADE

from splade.splade.models.transformer_rep import Splade
from transformers import AutoTokenizer

splade_model_id = 'naver/splade-cocondenser-ensembledistil'

splade_model = Splade(splade_model_id, agg='max')
splade_model.to('cpu')
splade_model.eval()

splade_tokenizer = AutoTokenizer.from_pretrained(splade_model_id)

다양한 search methods를 구현하기 위한 function들

from search_utils import fashion_query_transformer, clothes_detector, text_search, gen_sparse_vector, describe_clothes, additional_search

두 단계 search를 위한 local_DB

local_db = pd.read_csv("../data/imaterialist-fashion-2020-fgvc7/local_db.csv")
local_db['values'] = local_db['values'].apply(json.loads)

image detection을 위한 YOLO

from yolo_utils import fix_channels, visualize_predictions, rescale_bboxes, plot_results, box_cxcywh_to_xyxy
from transformers import YolosFeatureExtractor, YolosForObjectDetection

MODEL_NAME = "valentinafeve/yolos-fashionpedia"

yolo_feature_extractor = YolosFeatureExtractor.from_pretrained('hustvl/yolos-small')
yolo_model = YolosForObjectDetection.from_pretrained(MODEL_NAME)

# 미리 선정된 prediction labels
cats = ['shirt, blouse', 'top, t-shirt, sweatshirt', 'sweater', 'cardigan', 'jacket', 'vest', 'pants', 'shorts', 'skirt', 'coat', 'dress', 'jumpsuit', 'cape', 'glasses', 'hat', 'headband, head covering, hair accessory', 'tie', 'glove', 'watch', 'belt', 'leg warmer', 'tights, stockings', 'sock', 'shoe', 'bag, wallet', 'scarf', 'umbrella', 'hood', 'collar', 'lapel', 'epaulette', 'sleeve', 'pocket', 'neckline', 'buckle', 'zipper', 'applique', 'bead', 'bow', 'flower', 'fringe', 'ribbon', 'rivet', 'ruffle', 'sequin', 'tassel']

# 사용자 인풋 정의

from search_utils import fashion_query_transformer, clothes_detector, get_top_indices

# 1. Text input only

text_input = "a black cat"

# gateway
text_query = fashion_query_transformer(text_input)
print("### Result from the text_input gateway : \n{}".format(text_query))

if text_query:
    print("Searching ...")
    # text search
    result = text_search(index, text_query, clip_model, clip_tokenizer, splade_model, splade_tokenizer, top_k=10, hybrid=True)

    # 이미지들의 path 가져오기
    paths = dict()
    for k,v in result.items():
        paths[k] = ['../data/'+i['metadata']['img_path'] for i in v['matches']]

    # 이미지들 show
    for k,v in paths.items():
        print(k)
        draw_images([Image.open(i) for i in v])
else:
    print("패션과 무관한 텍스트 입니다. 다시 입력하세요.")


# 2. Image input only

image_path = "../data/imaterialist-fashion-2020-fgvc7/test_images/test_image7.jpg"

image = Image.open(image_path)

image = fix_channels(ToTensor()(image))
# object detections
print("Detecting items from the image.")
cropped_images = clothes_detector(image, yolo_feature_extractor, yolo_model, thresh=0.7)

if len(cropped_images.keys())==0:
    print("Nothing detected from the image")
else:
    print("Detected ", cropped_images.keys())
    
    # describe the labels I have found
    descriptions = dict()

    print("Start creating descriptions for each item.")
    for i, img in cropped_images.items():
        print("Created descriptions for {}".format(i))
        desc = describe_clothes(img, i, openai.api_key)
        descriptions[i] = desc
    print("\nTransform the descriptions into structured query.")
    text_query = fashion_query_transformer(str(descriptions))
    print(text_query)
    results = text_search(index, text_query, clip_model, clip_tokenizer, splade_model, splade_tokenizer, top_k=100)
    print("\nRetrieved 100 images based on text search")

    print("\nConducting additional search using the input images")

    results2 = additional_search(local_db, cropped_images, results, clip_processor, clip_model, clip_tokenizer, 10)

    for k,v in results2.items():
        print(k)
        draw_images([Image.open('../data/'+i) for i in v])

image

# 3. Text and Image input

- 기존의 이미지와 지나치게 다른 fashion style을 입력시, 원하는 결과가 나오지 않을 가능성이 높음

text_input = "softer and more comfortable material"
image_path = "../data/imaterialist-fashion-2020-fgvc7/test_images/test_image2.jpg"

# image

text_query = fashion_query_transformer(text_input)
print("### Result from the text_input gateway : \n{}".format(text_query))

# 패션과 유관한 쿼리

if 'clothes_type' in text_query['items'][0].keys():
    print("구체적인 아이템 보다는, 원하는 패션 스타일을 입력해주세요")
elif text_query:
    image = Image.open(image_path)
    image = fix_channels(ToTensor()(image))
    # object detections
    print("Detecting items from the image.")
    cropped_images = clothes_detector(image, yolo_feature_extractor, yolo_model)

    if len(cropped_images.keys())==0:
        print("Nothing detected from the image")
    else:
        print("Detected ", cropped_images.keys())
        print("-"*10, "Start image only search", "-"*10)
        
        # describe the labels I have found
        descriptions = dict()

        print("Start creating descriptions for each item.")
        for i, img in cropped_images.items():
            print("Created descriptions for {}".format(i))
            desc = describe_clothes(img, i, openai.api_key)
            descriptions[i] = desc
        print("\nTransform the descriptions into structured query.")
        text_query = fashion_query_transformer(str(descriptions))
        print(text_query)
        results = text_search(index, text_query, clip_model, clip_tokenizer, splade_model, splade_tokenizer, top_k=200)
        print("\nRetrieved 200 images based on text search")

        print("\nConducting additional search using the input images")

        results2 = additional_search(local_db, cropped_images, results, clip_processor, clip_model, clip_tokenizer, 100)
        print("\nRetrieved 100 items each, from sequential image search")

        print("-"*10, "Start reranking the results based on user input text", "-"*10)
        # 텍스트 검색
        new_results = list()

        for k,v in results2.items():
            ids = [os.path.splitext(os.path.basename(i))[0] for i in v]
            tmp = local_db.loc[local_db['vdb_id'].isin(ids)]

            r = get_top_indices(tmp, text_query['items'][0]['refined_text'], k, clip_processor, clip_model, clip_tokenizer, 10, type='text')
            new_results.append(r)

        refined_result = dict()

        for search_result in new_results:
            category = list(search_result.keys())[0]
            paths = list(search_result.values())[0]

            full_paths = [os.path.join("../data/imaterialist-fashion-2020-fgvc7", "cropped_images", i+".jpg") for i in paths]
            refined_result[category] = full_paths
        for k,v in refined_result.items():
            print(k)
            draw_images([Image.open('../data/'+i) for i in v])
else:
    print("패션과 무관한 텍스트 입니다.")
